+++ src/test/kotlin/racingcar/service/RacingServiceTest.kt
@@ -6,7 +6,7 @@ import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.MethodSource
 import org.junit.jupiter.params.provider.ValueSource
 import racingcar.domain.CarName
-import racingcar.model.Car
+import racingcar.domain.Car
 import java.util.stream.Stream
 import kotlin.test.assertEquals
